pyproject.toml
[tool.poetry]
name = "milestone4-python-bca"
version = "0.1.0"
description = ""
authors = ["samuelindrawijaya <samuelindraw75@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
flask = "^3.0.3"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.36"
pymysql = "^1.1.1"
flask-sqlalchemy = "^3.1.1"
flask-migrate = "^4.0.7"
flask-bcrypt = "^1.0.1"
flask-jwt-extended = "^4.6.0"
mysqlclient = "^2.2.5"
flask-login = "^0.6.3"
flasgger = "^0.9.7.1"
pyotp = "^2.9.0"
qrcode = {extras = ["pil"], version = "^8.0"}
werkzeug = "^3.0.6"
email-validator = "^2.2.0"
flask-mail = "^0.10.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
